refactor(conversion): use declarative converters
