+++ configio/internal/conversion/representation.go
@@ -1,14 +1,7 @@
 package conversion
 
 import (
-	"bytes"
-	"errors"
 	"fmt"
-	"io"
-	"net/http"
-	"net/url"
-	"strconv"
-	"time"
 
 	"github.com/benchttp/sdk/benchttp"
 )
@@ -56,198 +49,9 @@ func (repr Repr) Validate() error {
 // and stores any non-nil field value into the corresponding field
 // of dst.
 func (repr Repr) Decode(dst *benchttp.Runner) error {
-	if err := repr.decodeRequestField(dst); err != nil {
-		return err
-	}
-	if err := repr.decodeRunnerFields(dst); err != nil {
-		return err
-	}
-	return repr.decodeTestsField(dst)
-}
-
-func (repr Repr) decodeRequestField(dst *benchttp.Runner) error {
-	if dst.Request == nil {
-		dst.Request = &http.Request{}
-	}
-
-	if method := repr.Request.Method; method != nil {
-		dst.Request.Method = *method
-	}
-
-	if rawURL := repr.Request.URL; rawURL != nil {
-		parsedURL, err := parseAndBuildURL(*rawURL, repr.Request.QueryParams)
-		if err != nil {
-			return fmt.Errorf(`configio: invalid url: %q`, *rawURL)
-		}
-		dst.Request.URL = parsedURL
-	}
-
-	if header := repr.Request.Header; len(header) != 0 {
-		httpHeader := http.Header{}
-		for key, val := range header {
-			httpHeader[key] = val
-		}
-		dst.Request.Header = httpHeader
-	}
-
-	if body := repr.Request.Body; body != nil {
-		switch body.Type {
-		case "raw":
-			dst.Request.Body = io.NopCloser(bytes.NewReader([]byte(body.Content)))
-		default:
-			return errors.New(`configio: request.body.type: only "raw" accepted`)
-		}
-	}
-
-	return nil
-}
-
-func (repr Repr) decodeRunnerFields(dst *benchttp.Runner) error {
-	if requests := repr.Runner.Requests; requests != nil {
-		dst.Requests = *requests
-	}
-
-	if concurrency := repr.Runner.Concurrency; concurrency != nil {
-		dst.Concurrency = *concurrency
-	}
-
-	if interval := repr.Runner.Interval; interval != nil {
-		parsedInterval, err := parseOptionalDuration(*interval)
-		if err != nil {
+	for _, decoder := range converters {
+		if err := decoder.decode(repr, dst); err != nil {
 			return err
-		}
-		dst.Interval = parsedInterval
-	}
-
-	if requestTimeout := repr.Runner.RequestTimeout; requestTimeout != nil {
-		parsedTimeout, err := parseOptionalDuration(*requestTimeout)
-		if err != nil {
-			return err
-		}
-		dst.RequestTimeout = parsedTimeout
-	}
-
-	if globalTimeout := repr.Runner.GlobalTimeout; globalTimeout != nil {
-		parsedGlobalTimeout, err := parseOptionalDuration(*globalTimeout)
-		if err != nil {
-			return err
-		}
-		dst.GlobalTimeout = parsedGlobalTimeout
-	}
-
-	return nil
-}
-
-func (repr Repr) decodeTestsField(dst *benchttp.Runner) error {
-	testSuite := repr.Tests
-	if len(testSuite) == 0 {
-		return nil
-	}
-
-	cases := make([]benchttp.TestCase, len(testSuite))
-	for i, t := range testSuite {
-		fieldPath := func(caseField string) string {
-			return fmt.Sprintf("tests[%d].%s", i, caseField)
-		}
-
-		if err := requireConfigFields(map[string]interface{}{
-			fieldPath("name"):      t.Name,
-			fieldPath("field"):     t.Field,
-			fieldPath("predicate"): t.Predicate,
-			fieldPath("target"):    t.Target,
-		}); err != nil {
-			return err
-		}
-
-		field := benchttp.MetricsField(*t.Field)
-		if err := field.Validate(); err != nil {
-			return fmt.Errorf("%s: %s", fieldPath("field"), err)
-		}
-
-		predicate := benchttp.TestPredicate(*t.Predicate)
-		if err := predicate.Validate(); err != nil {
-			return fmt.Errorf("%s: %s", fieldPath("predicate"), err)
-		}
-
-		target, err := parseMetricValue(field, fmt.Sprint(t.Target))
-		if err != nil {
-			return fmt.Errorf("%s: %s", fieldPath("target"), err)
-		}
-
-		cases[i] = benchttp.TestCase{
-			Name:      *t.Name,
-			Field:     field,
-			Predicate: predicate,
-			Target:    target,
-		}
-	}
-
-	dst.Tests = cases
-	return nil
-}
-
-// helpers
-
-// parseAndBuildURL parses a raw string as a *url.URL and adds any extra
-// query parameters. It returns the first non-nil error occurring in the
-// process.
-func parseAndBuildURL(raw string, qp map[string]string) (*url.URL, error) {
-	u, err := url.ParseRequestURI(raw)
-	if err != nil {
-		return nil, err
-	}
-
-	// retrieve url query, add extra params, re-attach to url
-	if qp != nil {
-		q := u.Query()
-		for k, v := range qp {
-			q.Add(k, v)
-		}
-		u.RawQuery = q.Encode()
-	}
-
-	return u, nil
-}
-
-// parseOptionalDuration parses the raw string as a time.Duration
-// and returns the parsed value or a non-nil error.
-// Contrary to time.ParseDuration, it does not return an error
-// if raw == "".
-func parseOptionalDuration(raw string) (time.Duration, error) {
-	if raw == "" {
-		return 0, nil
-	}
-	return time.ParseDuration(raw)
-}
-
-func parseMetricValue(
-	field benchttp.MetricsField,
-	inputValue string,
-) (benchttp.MetricsValue, error) {
-	fieldType := field.Type()
-	handleError := func(v interface{}, err error) (benchttp.MetricsValue, error) {
-		if err != nil {
-			return nil, fmt.Errorf(
-				"value %q is incompatible with field %s (want %s)",
-				inputValue, field, fieldType,
-			)
-		}
-		return v, nil
-	}
-	switch fieldType {
-	case "int":
-		return handleError(strconv.Atoi(inputValue))
-	case "time.Duration":
-		return handleError(time.ParseDuration(inputValue))
-	default:
-		return nil, fmt.Errorf("unknown field: %s", field)
-	}
-}
-
-func requireConfigFields(fields map[string]interface{}) error {
-	for name, value := range fields {
-		if value == nil {
-			return fmt.Errorf("%s: missing field", name)
 		}
 	}
 	return nil
